修改数据库

+++ src/main/kotlin/com/gyh/keepaccounts/service/FileService.kt
@@ -2,7 +2,6 @@ package com.gyh.keepaccounts.service
 
 import com.gyh.keepaccounts.model.User
 import com.gyh.keepaccounts.model.view.UserResponseInfo
-import com.qcloud.cos.model.ObjectMetadata
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import org.springframework.web.multipart.MultipartFile
@@ -36,7 +35,9 @@ class FileService {
     fun deleteFile(path: String) = COSClient.deleteFile(path)
 
     fun deleteAll(files: String) {
-        val objectNames = files.split(" ")
-        return COSClient.deleteFile(objectNames)
+        if (files.isNotBlank()) {
+            val objectNames = files.split(" ")
+            return COSClient.deleteFile(objectNames)
+        }
     }
 }
